Делит скрипт на модули
