+++ js/util.js
@@ -1,0 +1,33 @@
+const getRandomPositiveInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomPositiveFloat = (min, max, digits = 1) => {
+  const lower = Math.min(Math.abs(min), Math.abs(max));
+  const upper = Math.max(Math.abs(min), Math.abs(max));
+  const result = Math.random() * (upper - lower) + lower;
+  return +result.toFixed(digits);
+};
+
+const getRandomArrayElement = (array) => array[getRandomPositiveInteger(0, array.length - 1)];
+
+
+const getRandomLengthArray = (array) => {
+  const RandomLengthArray = [];
+  array.map((element) => {
+    if (Math.round(Math.random())) {
+      RandomLengthArray.push(element);
+    }
+  });
+  return RandomLengthArray;
+};
+
+export {
+  getRandomPositiveInteger,
+  getRandomPositiveFloat,
+  getRandomArrayElement,
+  getRandomLengthArray
+};
